Allow parameter to be passed to new windows
Allow parameter to be passed to a new windows. If the registered component is registered the old way, it simply use it as it. If it is registered as an object with a "url" and a "extra" key it will pass the "extra" as a query which will allow the loadBalancer.job to be more dynamic. Usefull if you want to instance multiple process of the same script.